--- packages/docs/src/pages/header.module.css
@@ -1,30 +1,0 @@
-.title {
-  font-size: 3em;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-    Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-  font-weight: 700;
-  line-height: 1.2;
-  word-break: break-word;
-}
-
-.row {
-  display: flex;
-  flex-direction: row;
-  text-align: left;
-}
-
-.youtubeIframe{
-  align-self: center;
-}
-
-@media screen and (max-width: 1200px) {
-  .row {
-    flex-direction: column;
-    text-align: center;
-  }
-
-  .tabContainer {
-    margin-top: 1rem;
-    justify-content: center;
-  }
-}
