--- packages/docs/src/pages/styles.module.css
@@ -1,111 +1,0 @@
-/* stylelint-disable docusaurus/copyright-header */
-
-/**
- * CSS files with the .module.css suffix will be treated as CSS modules
- * and scoped locally.
- */
-
-.heroBanner {
-  padding: 4rem 0;
-  text-align: center;
-  position: relative;
-  overflow: hidden;
-}
-
-@media screen and (max-width: 966px) {
-  .heroBanner {
-    padding: 2rem;
-  }
-}
-
-.buttons {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.features {
-  display: flex;
-  align-items: center;
-  padding: 2rem 0;
-  width: 100%;
-}
-
-.featureImage {
-  height: 200px;
-  width: 300px;
-}
-
-.container * {
-  color: white;
-}
-
-/* snippets part : */
-
-.snippets {
-  max-width: 450px;
-  margin-top: 40px;
-  margin-bottom: 30px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.snippetsTitle {
-  font-size: 1.3em;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-    Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-  font-weight: 700;
-  line-height: 1.4;
-  margin-bottom: 20px;
-}
-
-.snippet {
-  max-width: 400px;
-  padding: 12px 12px 10px 20px;
-  background-color: #111111;
-  margin-bottom: 10px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  position: relative;
-  border-radius: 0.4rem;
-}
-
-.snippetValue {
-  color: white;
-  font-size: 1.5em;
-  font-family: monospace;
-}
-
-.copySnippet {
-  color: white;
-  right: 15px;
-  padding: 2px 10px;
-  border-radius: 0.4rem;
-  position: absolute;
-  width: fit-content;
-}
-
-.copySnippet:hover {
-  cursor: pointer;
-  background-color: rgba(255, 255, 255, 0.15);
-}
-
-@media screen and (max-width: 1200px) {
-  .snippet {
-    margin: 0 auto;
-    margin-bottom: 3rem;
-  }
-}
-
-@media screen and (max-width: 600px) {
-  .heroBanner {
-    padding: 1rem;
-  }
-  .snippetValue {
-    font-size: 1.25em;
-  }
-  .copySnippet {
-    font-size: 0.75em;
-  }
-}
